+++ src/main/java/com/sprint/mission/discodeit/common/Role.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity;
+package com.sprint.mission.discodeit.common;
 
 public enum Role {
   ADMIN,
